AI_model/requirements.txt
zmq
pandas
numpy
neat-python
